% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gprofiler.R
\name{profile}
\alias{profile}
\title{Profile the C++ code executed in an expression.}
\usage{
profile(expr, filename = NULL)
}
\arguments{
\item{expr}{The expression to be profiled.}

\item{filename}{An optional filename to write profiling information to.}
}
\description{
This profiles an expression and visualizes the profile in a web browser.
}
\details{
This requires the libprofile and pprof tools available via:
\code{sudo apt-get install libgoogle-perftools-dev}
\code{go get github.com/google/pprof}
}

