\name{set_base_url}
\alias{set_base_url}
\title{Set the base URL.}
\usage{set_base_url(url)
}
\description{Set the base URL.}
\details{Set the base URL. Useful for overriding the default URL (http://biit.cs.ut.ee/gprofiler)
with the bleeding-edge beta or an archived version.}
\arguments{\item{url}{the base URL.}
}

