\name{set_user_agent}
\alias{set_user_agent}
\title{Set custom user agent string.}
\usage{set_user_agent(ua, append=F)
}
\description{Set custom user agent string.}
\details{Set the HTTP User-Agent string. Useful for overriding the default user agent for
packages that depend on gProfileR functionality.}
\arguments{\item{ua}{the user agent string.}
\item{append}{logical indicating whether to append the passed string to the default user agent string.}
}

